refactor: Move test cases
