+++ src/tests.rs
@@ -18,6 +18,7 @@ use std::time::Duration;
 /// Test cases are placed in the file of the most specific feature they test.
 /// With the following definition of specific (ordered by most specific first):
 ///
+///     Any key
 ///     Virtual modifiers
 ///     Disguised events input (i.e. transformation of relative event to pseudo keys)
 ///     Multipurpose keys (tap-preferred)
@@ -171,69 +172,6 @@ fn test_exact_match_default() {
             Action::KeyEvent(KeyEvent::new(Key::KEY_RIGHT, KeyValue::Release)),
             Action::Delay(Duration::from_nanos(0)),
             Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTALT, KeyValue::Press)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-        ],
-    )
-}
-
-#[test]
-fn test_exact_match_true_nested() {
-    assert_actions(
-        indoc! {"
-        keymap:
-          - exact_match: true
-            remap:
-              C-x:
-                remap:
-                  h: C-a
-        "},
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_LEFTSHIFT),
-            Event::key_press(Key::KEY_H),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTSHIFT, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_H, KeyValue::Press)),
-        ],
-    )
-}
-
-#[test]
-fn test_exact_match_false_nested() {
-    assert_actions(
-        indoc! {"
-        keymap:
-          - exact_match: false
-            remap:
-              C-x:
-                remap:
-                  h: C-a
-        "},
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_LEFTSHIFT),
-            Event::key_press(Key::KEY_H),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTSHIFT, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
             Action::Delay(Duration::from_nanos(0)),
             Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
         ],
@@ -417,185 +355,6 @@ fn test_device_override() {
 }
 
 #[test]
-fn test_merge_remaps() {
-    let config = indoc! {"
-        keymap:
-          - remap:
-              C-x:
-                remap:
-                  h: C-a
-          - remap:
-              C-x:
-                remap:
-                  k: C-w
-    "};
-
-    assert_actions(
-        config,
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_H),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-        ],
-    );
-
-    assert_actions(
-        config,
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_K),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_W, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_W, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-        ],
-    )
-}
-
-#[test]
-fn test_merge_remaps_with_override() {
-    let config = indoc! {"
-        keymap:
-          - remap:
-              C-x:
-                remap:
-                  h: C-a
-          - remap:
-              C-x:
-                remap:
-                  h: C-b
-                  c: C-q
-    "};
-
-    assert_actions(
-        config,
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_H),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-        ],
-    );
-
-    assert_actions(
-        config,
-        vec![
-            Event::key_press(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_X),
-            Event::key_release(Key::KEY_X),
-            Event::key_release(Key::KEY_LEFTCTRL),
-            Event::key_press(Key::KEY_C),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_X, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_Q, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_Q, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_LEFTCTRL, KeyValue::Release)),
-        ],
-    )
-}
-
-#[test]
-fn test_mixing_keypress_and_remap_in_keymap_action() {
-    // KEY_D will be emitted, and the remap will be used for next key press.
-    assert_actions(
-        indoc! {"
-        keymap:
-          - remap:
-              f12:
-                - d
-                - remap:
-                    a: b
-        "},
-        vec![
-            Event::key_press(Key::KEY_F12),
-            Event::key_release(Key::KEY_F12),
-            Event::key_press(Key::KEY_A),
-            Event::key_release(Key::KEY_A),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_D, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_D, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_F12, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_B, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_B, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-        ],
-    )
-}
-
-#[test]
-fn test_mixing_no_keypress_and_remap_in_keymap_action() {
-    // The first match stops the search for matches. So the last remap isn't used.
-    assert_actions(
-        indoc! {"
-        keymap:
-          - remap:
-              f12: []
-          - remap:
-              f12:
-                - remap:
-                    a: b
-        "},
-        vec![
-            Event::key_press(Key::KEY_F12),
-            Event::key_release(Key::KEY_F12),
-            Event::key_press(Key::KEY_A),
-            Event::key_release(Key::KEY_A),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_F12, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-        ],
-    )
-}
-
-#[test]
 fn test_no_keymap_action() {
     assert_actions(
         indoc! {"
@@ -620,57 +379,6 @@ fn test_no_keymap_action() {
         vec![Event::key_press(Key::KEY_F12), Event::key_release(Key::KEY_F12)],
         vec![Action::KeyEvent(KeyEvent::new(Key::KEY_F12, KeyValue::Release))],
     )
-}
-
-#[test]
-fn test_any_key() {
-    assert_actions(
-        indoc! {"
-        keymap:
-          - remap:
-              a: b
-              ANY: null
-        "},
-        vec![
-            Event::key_press(Key::KEY_A),
-            Event::key_release(Key::KEY_A),
-            Event::key_press(Key::KEY_C),
-            Event::key_release(Key::KEY_C),
-        ],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_B, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_B, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_A, KeyValue::Release)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_C, KeyValue::Release)),
-        ],
-    );
-}
-
-#[test]
-fn test_any_does_not_work_in_nested_remap() {
-    // This happens because it tries nested with K, and then cancels the nested
-    // remap because K is not mapped there. So it doesn't matter that it's later
-    // tried with any-key.
-    assert_actions(
-        indoc! {"
-        keymap:
-          - remap:
-              a:
-                - remap:
-                    b: c
-                    any: d
-              k: h
-        "},
-        vec![Event::key_press(Key::KEY_A), Event::key_press(Key::KEY_K)],
-        vec![
-            Action::KeyEvent(KeyEvent::new(Key::KEY_H, KeyValue::Press)),
-            Action::KeyEvent(KeyEvent::new(Key::KEY_H, KeyValue::Release)),
-            Action::Delay(Duration::from_nanos(0)),
-            Action::Delay(Duration::from_nanos(0)),
-        ],
-    );
 }
 
 #[test]
